Split up better tests and match bet tests into their own files
